+++ packages/morphic-fastcheck-interpreters/src/model/unknown.ts
@@ -1,10 +1,23 @@
 import type { AnyEnv } from '@morphic-ts/common/lib/config'
 import { memo } from '@morphic-ts/common/lib/utils'
 import type { ModelAlgebraUnknown } from '@morphic-ts/model-algebras/lib/unknown'
+import type { Arbitrary } from 'fast-check'
 import { anything } from 'fast-check'
+import { pipe } from 'fp-ts/function'
 
 import { fastCheckApplyConfig } from '../config'
 import { FastCheckType, FastCheckURI } from '../hkt'
+
+declare module '@morphic-ts/model-algebras/lib/unknown' {
+  /**
+   *  @since 0.0.1
+   */
+  export interface UnknownConfig {
+    [FastCheckURI]: {
+      arb: Arbitrary<unknown>
+    }
+  }
+}
 
 /**
  *  @since 0.0.1
@@ -12,6 +25,7 @@ import { FastCheckType, FastCheckURI } from '../hkt'
 export const fastCheckUnknownInterpreter = memo(
   <Env extends AnyEnv>(): ModelAlgebraUnknown<FastCheckURI, Env> => ({
     _F: FastCheckURI,
-    unknown: configs => env => new FastCheckType(fastCheckApplyConfig(configs)(anything(), env, {}))
+    unknown: configs => env =>
+      pipe(anything(), arb => new FastCheckType(fastCheckApplyConfig(configs)(arb, env, { arb })))
   })
 )