+++ packages/morphic-io-ts-interpreters/src/model/set.ts
@@ -1,10 +1,23 @@
 import type { AnyEnv } from '@morphic-ts/common/lib/config'
 import { memo } from '@morphic-ts/common/lib/utils'
 import type { ModelAlgebraSet } from '@morphic-ts/model-algebras/lib/set'
+import { pipe } from 'fp-ts/function'
+import type { Type } from 'io-ts'
 import { setFromArray } from 'io-ts-types/lib/setFromArray'
 
 import { IOTSType, IoTsURI } from '../hkt'
 import { iotsApplyConfig } from './../config'
+
+declare module '@morphic-ts/model-algebras/lib/set' {
+  /**
+   *  @since 0.0.1
+   */
+  export interface SetConfig<L, A> {
+    [IoTsURI]: {
+      type: Type<A, L>
+    }
+  }
+}
 
 /**
  *  @since 0.0.1
@@ -12,6 +25,7 @@ import { iotsApplyConfig } from './../config'
 export const ioTsSetInterpreter = memo(
   <Env extends AnyEnv>(): ModelAlgebraSet<IoTsURI, Env> => ({
     _F: IoTsURI,
-    set: (a, ord, config) => env => new IOTSType(iotsApplyConfig(config)(setFromArray(a(env).type, ord), env, {}))
+    set: (a, ord, config) => env =>
+      pipe(a(env).type, type => new IOTSType(iotsApplyConfig(config)(setFromArray(type, ord), env, { type })))
   })
 )