+++ packages/morphic-json-schema-interpreters/src/model/set.ts
@@ -4,9 +4,22 @@ import type { ModelAlgebraSet } from '@morphic-ts/model-algebras/lib/set'
 import { pipe } from 'fp-ts/pipeable'
 import { chainEitherK as SEchainEitherK } from 'fp-ts-contrib/lib/StateEither'
 
+import type { JsonSchemaResult } from '../hkt'
 import { JsonSchema, JsonSchemaURI } from '../hkt'
+import type { OptionalJSONSchema } from '../json-schema/json-schema-ctors'
 import { SetFromArrayTypeCtor } from '../json-schema/json-schema-ctors'
 import { jsonSchemaApplyConfig } from './../config'
+
+declare module '@morphic-ts/model-algebras/lib/set' {
+  /**
+   *  @since 0.0.1
+   */
+  export interface SetConfig<L, A> {
+    [JsonSchemaURI]: {
+      schema: JsonSchemaResult<OptionalJSONSchema>
+    }
+  }
+}
 
 /**
  *  @since 0.0.1
@@ -15,8 +28,12 @@ export const jsonSchemaSetInterpreter = memo(
   <Env extends AnyEnv>(): ModelAlgebraSet<JsonSchemaURI, Env> => ({
     _F: JsonSchemaURI,
     set: (getSchema, _name, config) => env =>
-      new JsonSchema(
-        jsonSchemaApplyConfig(config)(pipe(getSchema(env).schema, SEchainEitherK(SetFromArrayTypeCtor)), env, {})
+      pipe(
+        getSchema(env).schema,
+        schema =>
+          new JsonSchema(
+            jsonSchemaApplyConfig(config)(pipe(schema, SEchainEitherK(SetFromArrayTypeCtor)), env, { schema })
+          )
       )
   })
 )