+++ packages/morphic-model-algebras/src/str-map.ts
@@ -19,15 +19,25 @@ declare module '@morphic-ts/algebras/lib/hkt' {
 /**
  *  @since 0.0.1
  */
+export interface StrMapConfig<L, A> {}
+
+/**
+ *  @since 0.0.1
+ */
+export interface RecordConfig<LA, A, LB, B> {}
+
+/**
+ *  @since 0.0.1
+ */
 export interface ModelAlgebraStrMap<F extends URIS, Env extends AnyEnv> {
   _F: F
   strMap: <L, A>(
     codomain: Kind<F, Env, L, A>,
-    config?: ConfigsForType<Env, Readonly<Record<string, L>>, Readonly<Record<string, A>>>
+    config?: ConfigsForType<Env, Readonly<Record<string, L>>, Readonly<Record<string, A>>, StrMapConfig<L, A>>
   ) => Kind<F, Env, Readonly<Record<string, L>>, Readonly<Record<string, A>>>
   record: <LA extends string, LB, A extends string, B>(
     domain: Kind<F, Env, LA, A>,
     codomain: Kind<F, Env, LB, B>,
-    config?: ConfigsForType<Env, Readonly<Record<LA, LB>>, Readonly<Record<A, B>>>
+    config?: ConfigsForType<Env, Readonly<Record<LA, LB>>, Readonly<Record<A, B>>, RecordConfig<LA, A, LB, B>>
   ) => Kind<F, Env, Readonly<Record<LA, LB>>, Readonly<Record<A, B>>>
 }