+++ packages/morphic-ord-interpreters/src/model/newtype.ts
@@ -1,9 +1,9 @@
 import type { AnyEnv } from '@morphic-ts/common/lib/config'
 import { memo } from '@morphic-ts/common/lib/utils'
 import type { AnyNewtype, ModelAlgebraNewtype, NewtypeA } from '@morphic-ts/model-algebras/lib/newtype'
+import { pipe } from 'fp-ts/lib/function'
 import type { Ord } from 'fp-ts/lib/Ord'
 import { fromCompare } from 'fp-ts/lib/Ord'
-import { pipe } from 'fp-ts/lib/pipeable'
 
 import { ordApplyConfig } from '../config'
 import { OrdType, OrdURI } from '../hkt'